Move child components into sub components folder
