+++ src/components/MapAndTableControls/MapAndTableControls.jsx
@@ -1,12 +1,12 @@
 import PropTypes from 'prop-types'
 import L from 'leaflet'
 import styled, { css } from 'styled-components'
-import FilterIndicatorPill from './FilterIndicatorPill'
-import ViewToggle from './ViewToggle'
+import FilterIndicatorPill from './components/FilterIndicatorPill'
+import ViewToggle from './components/ViewToggle'
 import { mediaQueryTabletLandscapeOnly } from '../../styles/mediaQueries'
 import { ButtonSecondary } from '../generic'
-import zoomToSelectedSites from '../assets/zoom_to_selected_sites.svg'
-import zoomToFiltered from '../assets/zoom_to_filtered.svg'
+import zoomToSelectedSites from '../../assets/zoom_to_selected_sites.svg'
+import zoomToFiltered from '../../assets/zoom_to_filtered.svg'
 import useResponsive from '../../hooks/useResponsive'
 import { useFilterProjectsContext } from '../../context/FilterProjectsContext'
 import { URL_PARAMS } from '../../constants/constants'
